requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
xgboost
scikit-learn
python-dotenv
google-genai